[FEAT/#171] 플러그인 고도화
