+++ build-logic/convention/src/main/java/kr/genti/convention/config/CommonPluginConfig.kt
@@ -2,8 +2,9 @@ package kr.genti.convention.config
 
 import kr.genti.convention.extension.getLibrary
 import kr.genti.convention.extension.implementation
-import kr.genti.convention.plugin.AndroidHiltPlugin
-import kr.genti.convention.plugin.AndroidKotlinPlugin
+import kr.genti.convention.plugin.HiltPlugin
+import kr.genti.convention.plugin.KotlinPlugin
+import kr.genti.convention.plugin.TestPlugin
 import org.gradle.api.Project
 import org.gradle.api.artifacts.VersionCatalogsExtension
 import org.gradle.kotlin.dsl.apply
@@ -11,8 +12,9 @@ import org.gradle.kotlin.dsl.dependencies
 import org.gradle.kotlin.dsl.getByType
 
 fun Project.configureAndroidCommonPlugin() {
-    apply<AndroidKotlinPlugin>()
-    apply<AndroidHiltPlugin>()
+    apply<KotlinPlugin>()
+    apply<HiltPlugin>()
+    apply<TestPlugin>()
     with(plugins) {
         apply("kotlin-kapt")
         apply("kotlin-parcelize")
