+++ build-logic/convention/src/main/java/kr/genti/convention/plugin/HiltPlugin.kt
@@ -1,21 +1,23 @@
 package kr.genti.convention.plugin
 
-import kr.genti.convention.extension.getBundle
+import kr.genti.convention.extension.getLibrary
 import kr.genti.convention.extension.getVersionCatalog
 import kr.genti.convention.extension.implementation
+import kr.genti.convention.extension.kapt
 import org.gradle.api.Plugin
 import org.gradle.api.Project
 import org.gradle.kotlin.dsl.dependencies
 
-class AndroidKotlinPlugin : Plugin<Project> {
+class HiltPlugin : Plugin<Project> {
     override fun apply(target: Project) = with(target) {
         with(pluginManager) {
-            apply("kotlin-android")
+            apply("dagger.hilt.android.plugin")
         }
 
         val libs = extensions.getVersionCatalog()
         dependencies {
-            implementation(libs.getBundle("kotlin"))
+            implementation(libs.getLibrary("hilt"))
+            kapt(libs.getLibrary("hiltAndroidCompiler"))
         }
     }
 }
