+++ build-logic/convention/src/main/java/kr/genti/convention/plugin/KotlinPlugin.kt
@@ -1,23 +1,21 @@
 package kr.genti.convention.plugin
 
-import kr.genti.convention.extension.getLibrary
+import kr.genti.convention.extension.getBundle
 import kr.genti.convention.extension.getVersionCatalog
 import kr.genti.convention.extension.implementation
-import kr.genti.convention.extension.kapt
 import org.gradle.api.Plugin
 import org.gradle.api.Project
 import org.gradle.kotlin.dsl.dependencies
 
-class AndroidHiltPlugin : Plugin<Project> {
+class KotlinPlugin : Plugin<Project> {
     override fun apply(target: Project) = with(target) {
         with(pluginManager) {
-            apply("dagger.hilt.android.plugin")
+            apply("kotlin-android")
         }
 
         val libs = extensions.getVersionCatalog()
         dependencies {
-            implementation(libs.getLibrary("hilt"))
-            kapt(libs.getLibrary("hiltAndroidCompiler"))
+            implementation(libs.getBundle("kotlin"))
         }
     }
 }
